+++ NEWS.md
@@ -16,6 +16,7 @@
   unpenalized covariates and the penalized predictors.
 - Fix bug in handling of a factor variable with multiple levels in the set of
   penalized predictors.
+- Use the correct sigma term in the computation of the elpd for gaussian models.
 
 # hsstan 0.7 (1 May 2020)